+++ src/viser/client/src/App.tsx
@@ -24,7 +24,7 @@ import {
 } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
 
-// Local imports
+// Local imports.
 import { SynchronizedCameraControls } from "./CameraControls";
 import { SceneNodeThreeObject } from "./SceneTree";
 import { ViewerContext, ViewerContextContents } from "./ViewerContext";
@@ -35,9 +35,8 @@ import { WebsocketMessageProducer } from "./WebsocketInterface";
 import { Titlebar } from "./Titlebar";
 import { ViserModal } from "./Modal";
 import { useSceneTreeState } from "./SceneTreeState";
-import { useThrottledMessageSender } from "./WebsocketFunctions";
+import { useThrottledMessageSender } from "./WebsocketUtils";
 import { rayToViserCoords } from "./WorldTransformUtils";
-import { ndcFromPointerXy, opencvXyFromPointerXy } from "./ClickUtils";
 import { theme } from "./AppTheme";
 import { FrameSynchronizedMessageHandler } from "./MessageHandler";
 import { PlaybackFromFile } from "./FilePlayback";
@@ -49,8 +48,45 @@ import { VISER_VERSION } from "./VersionInfo";
 
 // ======= Utility functions =======
 
+/** Turn a click event into a normalized device coordinate (NDC) vector.
+ * Normalizes click coordinates to be between -1 and 1, with (0, 0) being the center of the screen.
+ *
+ * Returns null if input is not valid.
+ */
+function ndcFromPointerXy(
+  viewer: ViewerContextContents,
+  xy: [number, number],
+): THREE.Vector2 | null {
+  const mouseVector = new THREE.Vector2();
+  mouseVector.x =
+    2 * ((xy[0] + 0.5) / viewer.mutable.current.canvas!.clientWidth) - 1;
+  mouseVector.y =
+    1 - 2 * ((xy[1] + 0.5) / viewer.mutable.current.canvas!.clientHeight);
+  return mouseVector.x < 1 &&
+    mouseVector.x > -1 &&
+    mouseVector.y < 1 &&
+    mouseVector.y > -1
+    ? mouseVector
+    : null;
+}
+
+/** Turn a click event to normalized OpenCV coordinate (NDC) vector.
+ * Normalizes click coordinates to be between (0, 0) as upper-left corner,
+ * and (1, 1) as lower-right corner, with (0.5, 0.5) being the center of the screen.
+ * Uses offsetX/Y, and clientWidth/Height to get the coordinates.
+ */
+function opencvXyFromPointerXy(
+  viewer: ViewerContextContents,
+  xy: [number, number],
+): THREE.Vector2 {
+  const mouseVector = new THREE.Vector2();
+  mouseVector.x = (xy[0] + 0.5) / viewer.mutable.current.canvas!.clientWidth;
+  mouseVector.y = (xy[1] + 0.5) / viewer.mutable.current.canvas!.clientHeight;
+  return mouseVector;
+}
+
 /** Gets default WebSocket server URL based on current window location. */
-const getDefaultServerFromUrl = () => {
+const getDefaultServerFromUrl = (): string => {
   let server = window.location.href;
   server = server.replace("http://", "ws://");
   server = server.replace("https://", "wss://");
@@ -60,7 +96,7 @@ const getDefaultServerFromUrl = () => {
 };
 
 /** Disables rendering when component is not in view. */
-const DisableRender = () => useFrame(() => null, 1000);
+const DisableRender = (): null => useFrame(() => null, 1000);
 
 // ======= Main component tree =======
 
@@ -87,7 +123,7 @@ export function Root() {
     </div>
   );
 
-  // If dummy window dimensions are specified, wrap content in MacWindowWrapper
+  // If dummy window dimensions are specified, wrap content in MacWindowWrapper.
   if (!dummyWindowParam) return content;
 
   const [width, height] = dummyWindowParam.split("x").map(Number);
@@ -377,14 +413,14 @@ function ViewerCanvas({ children }: { children: React.ReactNode }) {
     if (ndcFromPointerXy(viewer, pointerXy) === null) return;
     pointerInfo.dragEnd = pointerXy;
 
-    // Check if pointer moved enough to be considered a drag
+    // Check if pointer moved enough to be considered a drag.
     if (
       Math.abs(pointerInfo.dragEnd[0] - pointerInfo.dragStart[0]) <= 3 &&
       Math.abs(pointerInfo.dragEnd[1] - pointerInfo.dragStart[1]) <= 3
     )
       return;
 
-    // Draw selection rectangle if in rect-select mode
+    // Draw selection rectangle if in rect-select mode.
     if (pointerInfo.enabled === "rect-select") {
       const ctx = mutable.current.canvas2d!.getContext("2d")!;
       ctx.clearRect(0, 0, ctx.canvas.width, ctx.canvas.height);
