+++ src/viser/client/src/CameraControls.tsx
@@ -7,7 +7,7 @@ import { useFrame } from "@react-three/fiber";
 import { PerspectiveCamera } from "three";
 import * as THREE from "three";
 import { computeT_threeworld_world } from "./WorldTransformUtils";
-import { useThrottledMessageSender } from "./WebsocketFunctions";
+import { useThrottledMessageSender } from "./WebsocketUtils";
 import { Grid, PivotControls } from "@react-three/drei";
 
 function OrbitOriginTool({
@@ -96,7 +96,7 @@ export function SynchronizedCameraControls() {
 
   const viewerMutable = viewer.mutable.current;
 
-  // Animation state interface
+  // Animation state interface.
   interface CameraAnimation {
     startUp: THREE.Vector3;
     targetUp: THREE.Vector3;
@@ -109,7 +109,7 @@ export function SynchronizedCameraControls() {
   const [cameraAnimation, setCameraAnimation] =
     useState<CameraAnimation | null>(null);
 
-  // Animation parameters
+  // Animation parameters.
   const ANIMATION_DURATION = 0.5; // seconds
 
   useFrame((state) => {
@@ -120,29 +120,29 @@ export function SynchronizedCameraControls() {
       const elapsed = state.clock.getElapsedTime() - cameraAnimation.startTime;
       const progress = Math.min(elapsed / cameraAnimation.duration, 1);
 
-      // Smooth step easing
+      // Smooth step easing.
       const t = progress * progress * (3 - 2 * progress);
 
-      // Interpolate up vector
+      // Interpolate up vector.
       const newUp = new THREE.Vector3()
         .copy(cameraAnimation.startUp)
         .lerp(cameraAnimation.targetUp, t)
         .normalize();
 
-      // Interpolate look-at position
+      // Interpolate look-at position.
       const newLookAt = new THREE.Vector3()
         .copy(cameraAnimation.startLookAt)
         .lerp(cameraAnimation.targetLookAt, t);
 
       camera.up.copy(newUp);
 
-      // Back up position
+      // Back up position.
       const prevPosition = new THREE.Vector3();
       cameraControls.getPosition(prevPosition);
 
       cameraControls.updateCameraUp();
 
-      // Restore position and set new look-at
+      // Restore position and set new look-at.
       cameraControls.setPosition(
         prevPosition.x,
         prevPosition.y,
@@ -160,7 +160,7 @@ export function SynchronizedCameraControls() {
         false,
       );
 
-      // Clear animation when complete
+      // Clear animation when complete.
       if (progress >= 1) {
         setCameraAnimation(null);
       }
@@ -178,13 +178,13 @@ export function SynchronizedCameraControls() {
     const cameraControls = viewerMutable.cameraControl;
     const camera = viewerMutable.cameraControl.camera;
 
-    // Get target up vector from matrix
+    // Get target up vector from matrix.
     const targetUp = new THREE.Vector3().setFromMatrixColumn(matrix, 1);
 
-    // Get current look-at position
+    // Get current look-at position.
     const currentLookAt = cameraControls.getTarget(new THREE.Vector3());
 
-    // Start new animation
+    // Start new animation.
     setCameraAnimation({
       startUp: camera.up.clone(),
       targetUp: targetUp,
@@ -218,15 +218,15 @@ export function SynchronizedCameraControls() {
       .normalize();
     const angle = Math.acos(Math.min(1, Math.max(-1, cameraUp.dot(pivotUp))));
 
-    // Create rotation matrix
+    // Create rotation matrix.
     const rotationMatrix = new THREE.Matrix4();
     if (axis.lengthSq() > 0.0001) {
-      // Check if cross product is valid
+      // Check if cross product is valid.
       rotationMatrix.makeRotationAxis(axis, angle);
     }
     // rotationMatrix.premultiply(origRotation);
 
-    // Combine rotation with position
+    // Combine rotation with position.
     const matrix = new THREE.Matrix4();
     matrix.multiply(rotationMatrix);
     matrix.multiply(origRotation);
@@ -294,7 +294,7 @@ export function SynchronizedCameraControls() {
     camera_control.getTarget(lookAt).applyQuaternion(R_world_threeworld);
     const up = three_camera.up.clone().applyQuaternion(R_world_threeworld);
 
-    //Store initial camera values
+    // Store initial camera values.
     if (initialCameraRef.current === null) {
       initialCameraRef.current = {
         camera: three_camera.clone(),
