+++ src/viser/client/src/components/Folder.css.ts
@@ -2,17 +2,15 @@ import { style } from "@vanilla-extract/css";
 import { vars } from "../AppTheme";
 
 export const folderWrapper = style({
+  borderWidth: "1px",
   position: "relative",
   marginLeft: vars.spacing.xs,
   marginRight: vars.spacing.xs,
   // If there's a GUI element above, we need more margin.
-  marginTop: vars.spacing.md,
-  ":first-child": {
-    marginTop: vars.spacing.xs,
-  },
+  marginTop: vars.spacing.xs,
   // If there's a GUI element below, we need more margin.
   // Note: 0.5em is the vertical margin below general GUI elements.
-  marginBottom: "0.7em",
+  marginBottom: "1.2em",
   ":last-child": {
     marginBottom: "0.5em",
   },
