+++ src/viser/client/src/components/Html.tsx
@@ -1,7 +1,7 @@
 import { GuiHtmlMessage } from "../WebsocketMessages";
 
 function HtmlComponent({ props }: GuiHtmlMessage) {
-  if (!props.visible) return <></>;
+  if (!props.visible) return null;
   return <div dangerouslySetInnerHTML={{ __html: props.content }} />;
 }
