+++ src/viser/client/src/components/Vector3.tsx
@@ -9,7 +9,7 @@ export default function Vector3Component({
   props: { hint, label, visible, disabled, min, max, step, precision },
 }: GuiVector3Message) {
   const { setValue } = React.useContext(GuiComponentContext)!;
-  if (!visible) return <></>;
+  if (!visible) return null;
   return (
     <ViserInputComponent {...{ uuid, hint, label }}>
       <VectorInput
