+++ src/viser/client/src/mesh/BatchedMeshHoverOutlines.tsx
@@ -18,7 +18,7 @@ interface BatchedMeshHoverOutlinesProps {
     rotation: THREE.Quaternion;
     scale: THREE.Vector3;
   };
-  // Function to compute batch index from instance index - needed for cases like InstancedAxes
+  // Function to compute batch index from instance index - needed for cases like InstancedAxes.
   // where instanceId (from hover) doesn't match batched_positions indexing
   computeBatchIndexFromInstanceIndex?: (instanceId: number) => number;
 }
@@ -47,33 +47,33 @@ export const BatchedMeshHoverOutlines: React.FC<
   meshTransform,
   computeBatchIndexFromInstanceIndex,
 }) => {
-  // Get hover state from context
+  // Get hover state from context.
   const hoveredRef = React.useContext(HoverableContext)!;
 
-  // Create outline mesh reference
+  // Create outline mesh reference.
   const outlineRef = React.useRef<THREE.Mesh>(null);
 
-  // Get rendering context for screen size
+  // Get rendering context for screen size.
   const gl = useThree((state) => state.gl);
   const contextSize = React.useMemo(
     () => gl.getDrawingBufferSize(new THREE.Vector2()),
     [gl],
   );
 
-  // Create outline geometry based on the original geometry using memoization
+  // Create outline geometry based on the original geometry using memoization.
   const outlineGeometry = React.useMemo(() => {
     if (!geometry) return null;
-    // Clone the geometry to create an independent copy for the outline
+    // Clone the geometry to create an independent copy for the outline.
     return geometry.clone();
   }, [geometry]);
 
-  // Create outline material with fixed properties
+  // Create outline material with fixed properties.
   const outlineMaterial = React.useMemo(() => {
     const material = new OutlinesMaterial({
       side: THREE.BackSide,
     });
 
-    // Set fixed properties to match OutlinesIfHovered
+    // Set fixed properties to match OutlinesIfHovered.
     material.thickness = 10;
     material.color = new THREE.Color(0xfbff00); // Yellow highlight color
     material.opacity = 0.8;
@@ -85,8 +85,8 @@ export const BatchedMeshHoverOutlines: React.FC<
     return material;
   }, [contextSize]);
 
-  // Separate cleanup for geometry and material to handle dependency changes correctly
-  // Clean up geometry when it changes or component unmounts
+  // Separate cleanup for geometry and material to handle dependency changes correctly.
+  // Clean up geometry when it changes or component unmounts.
   React.useEffect(() => {
     return () => {
       if (outlineGeometry) {
@@ -95,7 +95,7 @@ export const BatchedMeshHoverOutlines: React.FC<
     };
   }, [outlineGeometry]);
 
-  // Clean up material when it changes or component unmounts
+  // Clean up material when it changes or component unmounts.
   React.useEffect(() => {
     return () => {
       if (outlineMaterial) {
@@ -104,27 +104,27 @@ export const BatchedMeshHoverOutlines: React.FC<
     };
   }, [outlineMaterial]);
 
-  // Update outline position based on hover state
+  // Update outline position based on hover state.
   useFrame(() => {
     if (!outlineRef.current || !outlineGeometry || !hoveredRef) return;
 
-    // Hide by default
+    // Hide by default.
     outlineRef.current.visible = false;
 
-    // Check if we're hovering and have a valid instanceId
+    // Check if we're hovering and have a valid instanceId.
     if (
       hoveredRef.current.isHovered &&
       hoveredRef.current.instanceId !== null
     ) {
-      // Get the instance ID from the hover state
+      // Get the instance ID from the hover state.
       const hoveredInstanceId = hoveredRef.current.instanceId;
 
-      // Calculate the actual batch index using the mapping function if provided
+      // Calculate the actual batch index using the mapping function if provided.
       const batchIndex = computeBatchIndexFromInstanceIndex
         ? computeBatchIndexFromInstanceIndex(hoveredInstanceId)
         : hoveredInstanceId; // Default is identity mapping
 
-      // Create DataViews to read float values
+      // Create DataViews to read float values.
       const positionsView = new DataView(
         batched_positions.buffer,
         batched_positions.byteOffset,
@@ -143,14 +143,14 @@ export const BatchedMeshHoverOutlines: React.FC<
         const posOffset = batchIndex * 3 * 4; // 3 floats, 4 bytes per float
         const wxyzOffset = batchIndex * 4 * 4; // 4 floats, 4 bytes per float
 
-        // Position the outline at the hovered instance
+        // Position the outline at the hovered instance.
         outlineRef.current.position.set(
           positionsView.getFloat32(posOffset, true), // x
           positionsView.getFloat32(posOffset + 4, true), // y
           positionsView.getFloat32(posOffset + 8, true), // z
         );
 
-        // Set rotation to match the hovered instance
+        // Set rotation to match the hovered instance.
         outlineRef.current.quaternion.set(
           wxyzsView.getFloat32(wxyzOffset + 4, true), // x
           wxyzsView.getFloat32(wxyzOffset + 8, true), // y
@@ -198,7 +198,7 @@ export const BatchedMeshHoverOutlines: React.FC<
     }
   });
 
-  // This is now handled by the earlier cleanup effect
+  // This is now handled by the earlier cleanup effect.
 
   if (!hoveredRef || !outlineGeometry) {
     return null;
